remove idx-lite package, export OktaAuthIdxLite factory
