--- lib/exports/cdn/idx-lite.ts
@@ -1,4 +1,0 @@
-import { OktaAuth } from '../idx-lite';
-
-// Export only a single object
-export default OktaAuth;
