+++ lib/idx/factory/index.ts
@@ -1,2 +1,3 @@
 export * from './api';
 export * from './OktaAuthIdx';
+export * from './OktaAuthIdxLite';
