+++ rollup.config.js
@@ -15,7 +15,6 @@ let entries = {
   'okta-auth-js': 'lib/exports/default.ts',
   'core': 'lib/exports/core.ts',
   'authn': 'lib/exports/authn.ts',
-  'idx-lite': 'lib/exports/idx-lite.ts',
   'idx': 'lib/exports/idx.ts',
   'myaccount': 'lib/exports/myaccount.ts'
 };
